+++ apps/frontend/src/app/[locale]/lems/(volunteer)/(dashboard)/reports/judging-schedule/graphql/index.ts
@@ -1,0 +1,12 @@
+export { GET_JUDGING_SCHEDULE, parseJudgingSchedule } from './query';
+export type {
+  QueryData,
+  QueryVars,
+  Room,
+  Team,
+  AgendaEvent,
+  JudgingSession,
+  ScheduleRow
+} from './types';
+
+export * from './subscriptions';
