+++ apps/frontend/src/app/[locale]/lems/(volunteer)/(dashboard)/reports/judging-schedule/graphql/subscriptions/index.ts
@@ -1,0 +1,4 @@
+export {
+  TEAM_ARRIVAL_UPDATED_SUBSCRIPTION,
+  createTeamArrivalSubscription
+} from './team-arrival-updated';
